♻️ Refactor: strongly-type Gcal and GenericErrors
