+++ packages/backend/src/auth/controllers/auth.controller.ts
@@ -18,11 +18,9 @@ import compassAuthService from "@backend/auth/services/compass.auth.service";
 import GoogleAuthService, {
   getGAuthClientForUser,
 } from "@backend/auth/services/google.auth.service";
-import {
-  GcalError,
-  SyncError,
-} from "@backend/common/constants/error.constants";
+import { SyncError } from "@backend/common/constants/error.constants";
 import { error } from "@backend/common/errors/handlers/error.handler";
+import { GcalError } from "@backend/common/errors/integration/gcal/gcal.errors";
 import { isInvalidGoogleToken } from "@backend/common/services/gcal/gcal.utils";
 import {
   ReqBody,
