+++ packages/backend/src/common/errors/sync/sync.errors.ts
@@ -1,0 +1,4 @@
+import { Status } from "@core/errors/status.codes";
+import { ErrorMetadata } from "@backend/common/types/error.types";
+
+interface SyncErrors {}
