+++ packages/backend/src/sync/util/sync.queries.ts
@@ -1,10 +1,8 @@
 import { Origin } from "@core/constants/core.constants";
 import { Schema_CalendarList } from "@core/types/calendar.types";
 import { Payload_Resource_Events, Resource_Sync } from "@core/types/sync.types";
-import {
-  GenericError,
-  SyncError,
-} from "@backend/common/constants/error.constants";
+import { SyncError } from "@backend/common/constants/error.constants";
+import { GenericError } from "@backend/common/errors/generic/generic.errors";
 import { error } from "@backend/common/errors/handlers/error.handler";
 import { getPrimaryGcalId } from "@backend/common/services/gcal/gcal.utils";
 import mongoService from "@backend/common/services/mongo.service";
